Fix Issue 48
The above row may still fall down briefly when first pressed but returns
to expected position right away, depending on how close to the top of
the list you're grabbing the item. Also this fixes a problem that the
target slot does not line up with the active item. The user experience
is much better.