Fixes problem with target slot not lining up
When an item was dragged to the bottom of the list then back up,
the target slot was lagging several slots behind the active item.
The problem occurred because scrolling is permitted beyond the actual
end of the list in order to be able to move a row to the end. However,
when moving back up, the scroll position is then too high. This fix
gradually resets the scroll position to the natural end of the list
so the offsets work correctly