Increase timeout for measuring layout on mount
This seems to have caused a race condition which would sometimes
cause layout to be sensitive to the list's position on screen.

Also temporarily reverse 2d0800d. This caused a "drop-down" problem
when selecting items near the end of the list that seems to be worse
than the problem it fixed.

Prevent dragging up beyond the start of the list.