Fix animation (#61)
* Allow row data to be simple values

* Fix Issue 48

The above row may still fall down briefly when first pressed but returns
to expected position right away, depending on how close to the top of
the list you're grabbing the item. Also this fixes a problem that the
target slot does not line up with the active item. The user experience
is much better.

* Fixes problem with target slot not lining up

When an item was dragged to the bottom of the list then back up,
the target slot was lagging several slots behind the active item.
The problem occurred because scrolling is permitted beyond the actual
end of the list in order to be able to move a row to the end. However,
when moving back up, the scroll position is then too high. This fix
gradually resets the scroll position to the natural end of the list
so the offsets work correctly

* Fix issue #51

Add advice to the README to enable native animation.
This, together with the previous 2 commits, goes a long
way to improving animation quality.

* Add Sortable, and example app

* Above item still dropping down on iOS

* UIManager no longer seems to be needed

* Fix case where the component is offset

If a fixed navbar was placed above this component,
tracking of the target was off by that many pixels.
Use a relative moveY.

* Fix regression with short list

* Fix rows dropping down when row height not integer

Change font size in example to show problem. Even rows will drop down.
Change comparison in checkTargetElement() to allow for floating point
cumulative errors.
Change names of some variables in checkTargetElement() to improve
self-documentation.

* Make link to Sortable relative.

Works on this branch and any branch going forward now.

* Increase timeout for measuring layout on mount

This seems to have caused a race condition which would sometimes
cause layout to be sensitive to the list's position on screen.

Also temporarily reverse 2d0800d. This caused a "drop-down" problem
when selecting items near the end of the list that seems to be worse
than the problem it fixed.

Prevent dragging up beyond the start of the list.

* A more robust way to measure layout

Make sure rendering is done before calling measure by calling
back from onLayout()

http://stackoverflow.com/questions/29828971/react-native-measure-a-view#comment65856238_29842652